backend/requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Task Queue
celery[redis]==5.3.4
redis==4.6.0  # Compatible with celery 5.3.4
flower==2.0.1

# File Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
openpyxl==3.1.2
pandas==2.1.3

# Email
# Note: For Windows Outlook support, install requirements-windows.txt
azure-identity==1.16.0
msgraph-sdk==1.2.0

# Utils
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
aiofiles==23.2.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
isort==5.12.0